consumer/requirements.txt
boto==2.49.0
boto3==1.12.31
botocore==1.15.31
docutils==0.15.2
jmespath==0.9.5
kafka-python==2.0.1
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
